chore: Update logo url.
Switch from files.edx.org to the logos.openedx.org url for the "Powered
by Open edX" logos.  We should still keep the old url working as many
old versions of Open edX still referer to it.

We're making this change so that for future versions of the edx-platform
the DNS for these logos lives alongside the rest of the Open edX
infrastructure managed by tCRIL.