fix: capa numerical response with scientic notation (#30010)
I found a potential location for a solution to the numericalinput bug. The problem stemmed from. In the convert markdown -> xml process, here, the checkIsNumeric function rejects the input if stringValue.match(/[a-z]/i) which says: does this value contain any letters, regardless of case? This rejection would cause uses of e and i in answers like the described problem to be string input problems. A simple fix would be to replace that match with regex like '/[a-df-hj-z]/i'.