fix: remove Jenkins dependency on script input
The script for finding order dependency test failures
was dependent on inputs from Jenkins logs which we no
longer use.

This quick fix was to just use a list of tests with the
right format. A future iteration might process a new type
of output, like that from (pytest -v).