feat: Get microfrontend URLs from site configuration
Since Maple release, Learning Micro-Frontend (MFE) becomes the default courseware experience, but Multisites have not been considered.
The Learning MFE URL cannot be conifgured per site because the value is not being retrieved through the site configuration
by using the site configuration helper, as it is being done in other points of the platform to obtain other values from the settings.
This commit changes this behaviour introducing the retrieval of Learning but also Account, Profile and Authn MFE URLs through the helper.