fix: edx-enterprise to 3.44.0 and edx-rbac to 1.7.0
Overrides the SystemWideEnterpriseUserRoleAssignment.get_assignments() method to return list of (role, context) assignments where the first item in the list corresponds to the currently active enterprise for the user.

This approach should allow us to remedy the problem without changing behavior of the consumers of JWT-stored roles (e.g. in ecommerce).

ENT-5700