refactor: move noauth rebind ModuleSystem argument to service
This removes the `rebind_noauth_module_to_user` argument from the
ModuleSystem constructor and moves it to a separate service called
"rebinder" in the class `RebindModuleService`. This is used in the
LTI module to bind calls recieved by its noauth endpoint to bind
the module the real_user.

The original function is defined as a local function and calls its
parent function recursively to perform rebinding. Due to the heavy
dependency of local variables, this new service is defined in the same
module to minimize imports.