fix: bump pinned pyfilesystem2 version
This commit fixes unpickling of exceptions that are based on `fs.error.ResourceError`,
which definitely happens if such an exception is throwed withing a Celery task
and, in turn, causes Celery worker to shut down.

See the details here: https://stackoverflow.com/a/41809333

The new version overrides `__reduce__` as suggested.