refactor: deprecate node_path attribute of ModuleSystem
The node_path attribute & constructor argument of the ModuleSyste is
deprecated without any replacement service or fallback as there doesn't
seem to be any core blocks using it.

The CAPA modules and the blocks also reference to the node_path
attribute, but are linked to the LoncapaSystem, which are not deprecated
in this commit.