refactor: remove `error_descriptor_class` and `NonStaffErrorBlock`
It's safe to remove this because non-staff [1] users cannot access [2]
an `ErrorBlock`. We were able to reproduce this with and without this commit
with the following results:
1. Staff users were seeing the `ErrorBlock`.
2. Non-staff users were getting an empty `<div class="vert-mod"></div>`.

In theory, error blocks should be hidden in the Learning MFE because of this
option [3]. However, when we manually set `hide_access_error_blocks` to
`False`, we kept getting identical results (with and without this commit), so
it looks that the removal `NonStaffErrorBlock` was just omitted at some point.

[1] https://github.com/openedx/edx-platform/blob/a4ec4c1b8ebeb3d7920a70d1d6f993fc9fdd1d4f/lms/djangoapps/courseware/access.py#L419-L436
[2] https://github.com/openedx/edx-platform/blob/a4ec4c1b8ebeb3d7920a70d1d6f993fc9fdd1d4f/lms/djangoapps/courseware/access.py#L150-L151
[3] https://github.com/openedx/edx-platform/blob/92ca176fde41fe2fb1827b51293bb96ae9a21b3f/lms/djangoapps/courseware/views/views.py#L1547-L1551