fix: delete ORA staff grader postman collection
This file was originally checked into
edx-platform so that other developers
could import it into Postman (https://postman.com)
in order to easily query the ORA staff grader API.

Unfortunately, it is humongous (1.3 MB).
This slows down edx-platform repo cloning
and adds lots of noise when grepping
through edx-platform. This cost is not justifiable
for a closed-source 3rd-party developer tool.