refactor!: remove safe_lxml import compatibility hack (#30671)
In the past, safe_lxml was located at ./common/lib/safe_lxml/safe_lxml,
and was imported as so:

  import safe_lxml

Now, safe_lxml is located at ./openedx/core/lib/safe_lxml, and is
imported as so:

  import openedx.core.lib.safe_lxml

We added in a temporary backwards-compatibility hack, in the form
of the module ./safe_lxml/__init__.py, to support the old import
format (with a warning written to the logs).

Enough time has passed that it we feel it is safe to
remove that compatibility hack. After this commit, imports in
the form `import safe_lxml` will raise an ImportError.