build: Unify constraint calculation in requirements files [WIP]
In this approach, all of the .in files are separate layers, with no
reference to each other or even to the constraints file. Instead,
`lock-all.in` references all of them, and when compiled to `lock-all.txt`
produces a global solution file. `use-lock.in` is a piece of indirection
that just says "use lock-all.txt as a constraints file"; each layer file
is then compiled *with* use-lock.in to produce dependency solutions that
are all constraint-compatible with each other and can be safely combined.

This should prevent the problems we've seen where one layer produces a
solution that is then incompatible with a later layer. It may also allow
for performance improvements, although that is untested.

Also:

- Move `pre-requirements` prereq to correct target
- Unroll for-loops in compile-requirements to make it easier to
  maintain

NOTE: This is not complete, and would need adjustments to the Django
version substitution and some other requirements-file references.