fix: allow inactive (email unverified users) to add and refund entitlements
The default behavior of SessionAuthentication is to deny inactive users.

Open edX considers a user inactive until they have verified their email.

Unfortunately, this means that users with unverified emails may not be
able to use functionality gated by SessionAuthentication.

In this case, this means a user is unable to add or refund entitlements
without email verification.

This commit switches to using SessionAuthenticationAllowInactiveUser so
inactive users will be able to use their courses.

This does not change any permissions as is_active is not checked for
JwtAuthentication.