feat: Update versions of webpack and sass-loader.
We needed to update this because the old version of sass-loader was
still requiring node-sass even though we have moved over `sass` as our
sass dependency.  But updating sass-loader required us to update webpack
and extract-webpack-lpugin.

The old sass loader was requiring a version of node-sass that was so old
that it had a dependency on node-gyp that was old enough that it only
worked with python2.

This was all breaking `npm install` when run from an ubuntu 22.04
machine.

Webpack Release Notes:
* https://github.com/webpack/webpack/releases?page=35
* https://github.com/webpack/webpack/releases/tag/v3.0.0

extract-text-webpack-plugin Changelog:
* https://github.com/webpack-contrib/extract-text-webpack-plugin/blob/master/CHANGELOG.md

sass-loader Changelog:
* https://github.com/webpack-contrib/sass-loader/blob/master/CHANGELOG.md